[PL-25854]: Refactoring
