+++ 120-ng-manager/modules/ngsettings/contracts/java/io/harness/ngsettings/beans/SettingRequestDTO.java
@@ -3,6 +3,7 @@ package io.harness.ngsettings.beans;
 import io.harness.data.validator.EntityIdentifier;
 import io.harness.ng.core.setting.SettingCategory;
 
+import io.harness.ng.core.setting.SettingUpdateType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import javax.validation.constraints.NotNull;
 import org.hibernate.validator.constraints.NotBlank;
@@ -12,5 +13,6 @@ public class SettingRequestDTO {
   @NotNull @NotBlank @Schema(description = SettingConstants.CATEGORY) SettingCategory category;
   @NotNull @NotBlank @Schema(description = SettingConstants.ALLOW_OVERRIDES) Boolean allowOverrides;
   @Schema(description = SettingConstants.VALUE) String value;
-  @NotNull @NotBlank @Schema(description = SettingConstants.UPDATE_TYPE) SettingUpdateType updateType;
+  @NotNull @NotBlank @Schema(description = SettingConstants.UPDATE_TYPE)
+  SettingUpdateType updateType;
 }
