+++ 970-ng-commons/src/main/java/io/harness/ng/core/setting/SettingSource.java
@@ -1,4 +1,4 @@
-package io.harness.ngsettings.beans;
+package io.harness.ng.core.setting;
 
 public enum SettingSource {
   ACCOUNT,
