+++ 970-ng-commons/src/main/java/io/harness/ng/core/setting/SettingUpdateType.java
@@ -1,3 +1,3 @@
-package io.harness.ngsettings.beans;
+package io.harness.ng.core.setting;
 
 public enum SettingUpdateType { UPDATE, RESTORE }
